refactor to util
