+++ packages/next/src/index.ts
@@ -1,8 +1,8 @@
-import fs from 'fs';
 import type { NextConfig } from 'next';
 import path from 'path';
 import semver from 'semver';
 import { getNextBuilder } from './builder.js';
+import { maybeInvalidateCacheOnSwcChange } from './swc-cache.js';
 
 /**
  * Default directories to scan for workflows and steps.
@@ -138,51 +138,13 @@ export function withWorkflow(
       !process.env.WORKFLOW_NEXT_PRIVATE_BUILT &&
       phase !== 'phase-production-server'
     ) {
-      // Check swc-plugin version and invalidate cache if changed
+      // Check swc-plugin build hash and invalidate cache if changed
       const distDir = path.resolve(
         process.cwd(),
         nextConfig.distDir || '.next'
       );
-      const cacheDir = path.join(distDir, 'cache');
-      const devCacheDir = path.join(distDir, 'dev', 'cache');
-      const workflowJsonPath = path.join(cacheDir, 'workflow.json');
-      const swcPluginBuildHash = require('@workflow/swc-plugin/build-hash.json')
-        .buildHash as string;
+      maybeInvalidateCacheOnSwcChange(distDir);
 
-      let shouldInvalidateCache = false;
-      try {
-        const existing = JSON.parse(fs.readFileSync(workflowJsonPath, 'utf-8'));
-        if (existing.swcPluginBuildHash !== swcPluginBuildHash) {
-          shouldInvalidateCache = true;
-        }
-      } catch {
-        // File doesn't exist or is invalid
-        shouldInvalidateCache = true;
-      }
-
-      if (shouldInvalidateCache) {
-        console.log('workflow transform upgraded, invalidating Next.js cache');
-        // Delete cache directories
-        const cacheDirs = [cacheDir, devCacheDir];
-        for (const dir of cacheDirs) {
-          if (fs.existsSync(dir)) {
-            fs.rmSync(dir, { recursive: true, force: true });
-          }
-        }
-      }
-
-      // Write workflow.json lazily on process exit
-      process.on('exit', () => {
-        try {
-          fs.mkdirSync(cacheDir, { recursive: true });
-          fs.writeFileSync(
-            workflowJsonPath,
-            JSON.stringify({ swcPluginBuildHash }, null, 2)
-          );
-        } catch {
-          // Ignore errors on exit
-        }
-      });
       const shouldWatch = process.env.NODE_ENV === 'development';
       const NextBuilder = await getNextBuilder();
       const workflowBuilder = new NextBuilder({
